yay it renders with timings (but input is still buggy?)

+++ src/config.rs
@@ -1,9 +1,9 @@
 pub struct AppConfig {
-    pub delay: i32,
+    pub delay: u64,
     pub run_once: bool,
 }
 
-pub fn create_config_from_args(delay: i32, run_once: bool) -> AppConfig {
+pub fn create_config_from_args(delay: u64, run_once: bool) -> AppConfig {
     // TODO: read a config file
 
     // re-create an AppConfig with the argument values and config file values
